% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster.R
\name{plot_raster}
\alias{plot_raster}
\title{Plot GDAL matrix.}
\usage{
plot_raster(filename, add = FALSE)
}
\arguments{
\item{filename}{source}

\item{add}{add to the plot}
}
\description{
Plot a matrix directly read via GDAL to match the current device size.
}
\details{
The data is read in at the resolution required by the current device, assuming it's a single
panel.  TODO: read off the space to plot into from the device, and
read only that window from the source at the right resolution.
}
\examples{
## this file takes 5s to read in total from NetCDF, 24 seconds via GDAL
filename <- "/rdsi/PRIVATE/raad/data_local/www.bodc.ac.uk/gebco/GRIDONE_2D.nc"
library(vapour)
## read in to match the device it's less than a second
plot_raster(filename)
#maps::map(add = TRUE)
## library(sf); plot(st_transform(sst_c, 4326), add = T)
}
