% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{raster_info}
\alias{raster_info}
\title{Raster IO}
\usage{
raster_info(pszFilename)
}
\arguments{
\item{pszFilename}{data source}
}
\description{
Raster IO
}
\examples{
f <- system.file("extdata", "sst.tif", package = "vapour")
}
