% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{raster_io}
\alias{raster_io}
\title{Raster IO}
\usage{
raster_io(filename, window)
}
\arguments{
\item{filename}{data source}

\item{window}{src_offset, src_dim, out_dim}
}
\description{
Raster IO
}
\examples{
f <- system.file("extdata", "sst.tif", package = "vapour")
}
