% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gdal_geometry.R
\name{read_gdal_geometry}
\alias{read_gdal_geometry}
\title{Read geometry from source as a blob vector.}
\usage{
read_gdal_geometry(dsource, layer = 0)
}
\arguments{
\item{dsource}{data source string}

\item{layer}{layer index (0)}
}
\description{
Read geometry from source as a blob vector.
}
\examples{
f <- system.file("extdata", "point.shp", package = "vapour")
tibble::tibble(wkb = read_gdal_geometry(f))
sfile <- system.file("shape/nc.shp", package="sf")
#vapour(sfile)
read_gdal_table(sfile)
}
