% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gdal_attributes.R
\name{read_gdal_table}
\alias{read_gdal_table}
\title{Read attributes from source as a data frame}
\usage{
read_gdal_table(dsource, layer = 0, sql = NULL)
}
\arguments{
\item{dsource}{data source string}

\item{layer}{layer index (0)}

\item{sql}{optional sql query executed against the data source}
}
\description{
See here for SQL details http://www.gdal.org/ogr_sql.html
}
\examples{
f <- system.file("extdata", "point.shp", package = "vapour")
read_gdal_table(f, sql = "SELECT * FROM point WHERE FID > 4 AND FID < 8")
file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
read_gdal_table(mvfile, sql = "SELECT NAME, PLAN_REF, SHAPE_AREA * 2 AS area FROM list_locality_postcode_meander_valley WHERE POSTCODE < 7304")
read_gdal_table(mvfile)
}
