% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sql_vapour}
\alias{sql_vapour}
\title{SQL layer read}
\usage{
sql_vapour(dsource, layer = 0L, sql = "")
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}
}
\description{
Read layer attributes after SQL select
}
\examples{
sfile <- system.file("shape/nc.shp", package="sf")
vapour(sfile)
pfile <- "inst/extdata/point.shp"
vapour(pfile)
}
