% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-simple-raster.R
\name{ssraster}
\alias{ssraster}
\title{Super simple raster}
\usage{
ssraster(x)
}
\arguments{
\item{x}{filename (or dsn, GDAL)}
}
\value{
ssraster
}
\description{
A very light class to hold a GDAL source, and a promise to
pull the data into a raster pull_ssraster
}
\examples{
ff <- system.file("extdata/sst.tif", package= "vapour")
ss <- ssraster(ff)
ss
pull_ssraster(ss, pulldim = c(36, 18), resample = "Lanczos")
## works up as well as down
pull_ssraster(ss, pulldim = c(100, 360), resample = "Lanczos")
pull_ssraster(ss, pulldim = c(100, 360), resample = "Gauss")
}
