% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour-package.R
\docType{data}
\name{tas_wkt}
\alias{tas_wkt}
\title{Example WKT coordinate reference system}
\description{
A Lambert Azimuthal Equal Area Well-Known-Text string for a region
centred on Tasmania.
}
\details{
Created from '+proj=laea +lon_0=147 +lat_0=-42 +datum=WGS84'.
For use in a future warping example.
}
