% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_format}
\alias{to_format}
\title{Test GDAL read}
\usage{
to_format(dsource, layer = 0L, format = "json")
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}
}
\description{
Simple pointless function to learn the GDAL API.
}
\details{
Microprocessors, databases, servers.
}
\examples{
sfile <- system.file("shape/nc.shp", package="sf")
vgeom(sfile)
pfile <- "inst/extdata/point.shp"
vgeom(pfile)
}
