% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R, R/vapour_input.R
\name{sds_info}
\alias{sds_info}
\alias{vapour-deprecated}
\alias{vapour_read_geometry_what}
\alias{vapour_read_feature_what}
\title{Functions deprecated from vapour}
\usage{
sds_info(x)

vapour_read_geometry_what(dsource, layer = 0L, sql = "",
  what = "geometry", textformat = "")

vapour_read_feature_what(dsource, layer = 0L, sql = "", what = "geometry",
  textformat = "json")
}
\description{
These will be removed in a future release.
}
