% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/vapour-package.r
\docType{package}
\name{vapour}
\alias{vapour}
\alias{vapour}
\alias{vapour-package}
\title{Test GDAL read}
\usage{
vapour(dsource, layer = 0L)
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}
}
\description{
Simple pointless function to learn the GDAL API.

vapour.
}
\details{
Microprocessors, databases, servers.
}
\examples{
sfile <- system.file("shape/nc.shp", package="sf")
vapour(sfile)
pfile <- "inst/extdata/point.shp"
vapour(pfile)
}
