% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vapour_layer_names}
\alias{vapour_layer_names}
\title{Layer names}
\usage{
vapour_layer_names(dsource, sql = "")
}
\arguments{
\item{dsource}{data source}

\item{sql}{sql}
}
\value{
character vector of layer names
}
\description{
Layer names
}
\examples{
file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
vapour_layers(mvfile)
}
