% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vapour_read_attributes}
\alias{vapour_read_attributes}
\title{Vector attributes read}
\usage{
vapour_read_attributes(dsource, layer = 0L, sql = "")
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}

\item{sql}{if not empty this is executed against the data source (layer will be ignored)}
}
\description{
Read layer attributes, optionally after SQL select
}
\examples{
file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
vapour_read_attributes(mvfile)
sq <- "SELECT * FROM list_locality_postcode_meander_valley WHERE FID < 5"
vapour_read_attributes(mvfile, sql = sq)
dsource <- "inst/extdata/tab/list_locality_postcode_meander_valley.tab"
.Call('_vapour_vapour_read_attributes', PACKAGE = 'vapour', dsource, layer, sql)

}
