% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vapour_read_extent}
\alias{vapour_read_extent}
\title{GDAL geometry extent}
\usage{
vapour_read_extent(dsource, layer = 0L, sql = "")
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}
}
\description{
Read a GDAL geometry summary as just the native bounding box, the four
numbers xmin, xmax, ymin, ymax in the usual simple convention.
}
\examples{
sfile <- system.file("shape/nc.shp", package="sf")
vapour_read_geometry_extent(sfile)
}
