% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vapour_read_feature_what}
\alias{vapour_read_feature_what}
\title{Read GDAL feature thing 'what'}
\usage{
vapour_read_feature_what(dsource, layer = 0L, sql = "", what = "geometry",
  textformat = "json")
}
\arguments{
\item{dsource}{data source}

\item{layer}{layer}

\item{sql}{sql}

\item{what}{what to read, "geometry", "text", "extent"}

\item{textformat}{indicate text output format, available are "json" (default), "gml", "kml", "wkt"}
}
\description{
Read GDAL geometry as blob, text, or numeric extent.
}
\details{
\code{vapour_read_feature_what} will read a feature in various ways, as binary WKB, various text formats, a numeric
extent, and for each an option SQL string will be evaluated against the data source before reading.
The extent is the native bounding box, the four numbers xmin, xmax, ymin, ymax.
}
\examples{
file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
#tib <- tibble::tibble(wkb = vapour_read_geometry(mvfile)) \%>\%
#  bind_cols(read_gdal_table(mvfile))
pfile <- system.file("extdata/point.shp", package = "vapour")
vapour_read_geometry(pfile)

file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
vapour_read_geometry_text(mvfile)
pfile <- system.file("extdata/point.shp", package = "vapour")
vapour_read_geometry_text(pfile)

file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
vapour_read_extent(mvfile)
}
