% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vapour_read_geometry_cpp}
\alias{vapour_read_geometry_cpp}
\title{vapour_read_geometry_cpp}
\usage{
vapour_read_geometry_cpp(dsource, layer = 0L, sql = "", what = "geometry",
  textformat = "json")
}
\description{
vapour_read_geometry_cpp
}
