% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_read_block.R
\name{vapour_read_raster_block}
\alias{vapour_read_raster_block}
\title{Read or write raster block}
\usage{
vapour_read_raster_block(dsource, offset, dimension, band = 1L)
}
\arguments{
\item{dsource}{file name to read from, or write to}

\item{offset}{position x,y to start writing (0-based, y-top)}

\item{dimension}{window size to read from, or write to}

\item{band}{which band to write to (1-based)}
}
\value{
for vapour_read_raster_block, a list with vector of data
}
\description{
Read a 'block' from raster.
}
\examples{
f <- system.file("extdata", "sst.tif", package = "vapour")
v <- vapour_read_raster_block(f, c(0L, 0L), dimension = c(2L, 3L), band = 1L)
}
