% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_vrt.R
\name{vapour_vector_vrt}
\alias{vapour_vector_vrt}
\title{Create vector VRT}
\usage{
vapour_vector_vrt(
  dsn,
  layer = NULL,
  crs = NULL,
  select = "*",
  geom = NULL,
  dialect = "sqlite",
  limit_n = 0,
  where = ""
)
}
\arguments{
\item{dsn}{data source}

\item{layer}{layer (name or number)  FIXME 0- OR 1- based?}

\item{crs}{optional targect projection}

\item{select}{select subquery e.g. 'val1, val2, ConvexHull(geom) AS geom'}

\item{geom}{must be provided when crs is used, and must match syntax in 'select' of geometry SQL (use 'AS')}

\item{dialect}{sql dialect (SQLITE or OGRSQL most important ones)}

\item{limit_n}{optional limit on number of output objects}

\item{where}{subquer/ies for WHERE clause}
}
\value{
character string vector data source name for GDAL
}
\description{
Vector ViRTual data source with options for layer, output projection, select query, SQL dialect, limit
on number of output objects.
}
\details{
When 'crs' is used the name of the 'geom' is required, and will be determined as the source
default name. Because we want to have dynamic SQL for geometry functions we must also indicate
this name. The best way is to use 'select' subqueries like 'ConvexHull(geom) AS geom', this way the
geom name hasn't changed - but for complete control set in the select subquery AND in the geom arg.
}
\examples{
dsn <- system.file("extdata/sst_c.gpkg", package = "vapour", mustWork = TRUE)
vrt <- vapour_vector_vrt(dsn, layer = "sst_c", select = "sst, ConvexHull(geom) as chull", 
   limit_n = 4, dialect = "sqlite", crs = "+proj=laea +lat_0=-90", debug = T, out_field = "chull")
}
