% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_vrt.R
\name{vapour_vrt}
\alias{vapour_vrt}
\title{Virtual raster}
\usage{
vapour_vrt(
  dsn,
  extent = NULL,
  projection = NULL,
  sds = 1L,
  bands = NULL,
  ...,
  relative_to_vrt = FALSE
)
}
\arguments{
\item{dsn}{data source name, filepath, url, database connection string, or VRT text}

\item{extent}{(optional) numeric extent, xmin,xmax,ymin,ymax}

\item{projection}{(optional) character string, projection string ("auth:code", proj4, or WKT, or anything understood by PROJ, see Details)}

\item{sds}{which subdataset to select from a source with more than one}

\item{bands}{(optional) which band/s to include from the source}

\item{...}{ignored}

\item{relative_to_vrt}{default \code{FALSE}, if \code{TRUE} input strings that identify as files on the system are left as-is (by default they are made absolute at the R level)}
}
\value{
VRT character string (for use by GDAL-capable tools, i.e. reading raster)
}
\description{
Simple VRT creation of a GDAL virtual raster.
}
\details{
Create a GDAL data source string (to be used like a filename) with various helpers.
VRT stands for 'ViRTual'.

An input string will be converted to a single subdataset, use 'sds' argument to select.

If 'extent', 'projection' is provided this is applied to override the source's extent and/or projection.
(These might be invalid, or missing, so we facilitate correcting this).

If 'bands' is provided this is used to select a set of bands (numbered from 1), which might be repeated, or
in any order.

Common inputs for \code{projection} are WKT variants, 'AUTH:CODE's e.g.
'EPSG:3031', the 'OGC:CRS84' for long,lat WGS84, 'ESRI:<code>' and other
authority variants, and datum names such as 'WGS84','NAD27' recognized by
PROJ itself.

See help for 'SetFromUserInput' in 'OGRSpatialReference', and 'proj_create_crs_to_crs'.

\href{https://proj.org/development/reference/functions.html#c.proj_create_crs_to_crs}{c.proj_create_crs_to_crs}

\href{https://proj.org/development/reference/functions.html#c.proj_create}{c.proj_create}

\href{https://gdal.org/doxygen/classOGRSpatialReference.html#aec3c6a49533fe457ddc763d699ff8796}{SetFromUserInput}
}
\examples{
tif <- system.file("extdata", "sst.tif", package = "vapour", mustWork = TRUE)
vapour_vrt(tif)

vapour_vrt(tif, bands = c(1, 1))

}
