% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_vsi_list.R
\name{vapour_vsi_list}
\alias{vapour_vsi_list}
\title{Read GDAL virtual source contents}
\usage{
vapour_vsi_list(dsource, ...)
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL) with virtual prefix, see Details}

\item{...}{ignored}
}
\value{
character vector listing of items
}
\description{
Obtain the names of available items in a virtual file source.
}
\details{
The \code{dsource} must begin with a valid form of the special \code{vsiPREFIX}, for details
see \href{https://gdal.org/user/virtual_file_systems.html}{GDAL Virtual File Systems}.

Note that the listing is not recursive, and so cannot be used for automation. One would
use this function interactively to determine a  useable \code{/vsiPREFIX/dsource} data
source string.
}
