% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-input.R
\name{vapour_warp_raster}
\alias{vapour_warp_raster}
\title{Raster input (warp)}
\usage{
vapour_warp_raster(x, band = 1L, geotransform = NULL,
  dimension = NULL, wkt = "", set_na = TRUE, source_wkt = NULL)
}
\arguments{
\item{x}{data source string (file name or URL or database connection string)}

\item{band}{index of band to read (1-based)}

\item{geotransform}{the affine geotransform of the warped raster}

\item{dimension}{dimensions in pixels of the warped raster (x, y)}

\item{wkt}{projection of warped raster in Well-Known-Text}

\item{set_na}{NOT IMPLEMENTED logical, should 'NODATA' values be set to \code{NA}}

\item{source_wkt}{optional, override the projection of the source with WKT}
}
\value{
list of vectors (only 1 for 'band') of numeric values, in raster order
}
\description{
Read a window of data from a GDAL raster source through a warp specification.
Only a single band may be read. The warp specification is provided by 'geotransform',
'dimension', and 'wkt' properties of the transformed output.
}
\details{
This function is not memory safe, the source is left on disk but the output
raster is all computed in memory so please be careful with very large values
for 'dimension'. \code{1000 * 1000 * 8} for 1000 columns, 1000 rows and floating point
double type will be 8Mb.

There's no control over the output type (always double floating point) and no
control for the resampling algorithm (uses 'nearest neighbour').

'wkt' refers to the full Well-Known-Text specification of a coordinate reference
system. See \code{\link[=vapour_srs_wkt]{vapour_srs_wkt()}} for conversion from PROJ.4 string to WKT.

'geotransform' is the six-figure affine transform 'xmin, xres, yskew, ymax, xskew, yres' see \code{\link[=vapour_raster_info]{vapour_raster_info()}} for more detail.

'dimension' is the pixel dimensions of the output, x (ncol) then y (nrow).

Values for missing data are not yet handled, just returned
as-is. Note that there may be regions of "zero data" in a warped output,
separate from propagated missing "NODATA" values in the source.

Argument 'source_wkt' may be used to assign the proejction of the source, this
is sometimes required especially for NetCDF files (no checking is done). If the
source has no projection and this is not set GDAL will error.
}
\examples{
gt <- c(-637239.4, 5030.0, 0.0, 261208.7, 0.0, -7760.0)

f <- system.file("extdata", "sst.tif", package = "vapour")
vals <- vapour_warp_raster(f, geotransform = gt,
                              dimension = c(186, 298),
                              wkt = tas_wkt)
## wkt, dimension, geotransform above created via
##p <- raster::projectRaster(f,
## crs = "+proj=laea +lon_0=147 +lat_0=-42 +datum=WGS84")
##writeRaster(p, "a.tif")
## vapour::vapour_raster_info("a.tif")
}
