% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{raster_info}
\alias{raster_info}
\title{Raster information}
\usage{
raster_info(x)
}
\arguments{
\item{x}{data source string (i.e. file name or URL or database connection string)}
}
\description{
Return the basic structural metadata of a raster source understood by GDAL.
}
\details{
These are
\describe{
\item{geotransform}{the affine transform}
\item{dimXY}{dimensions x-y, columns*rows}
\item{minmax}{range of data values}
\item{tilesXY}{dimensions x-y of internal tiling scheme}
\item{projection}{text version of map projection parameter string}
}
The affine transform is the six parameter conversion of row-col position to map coordinates.
Usually the shear components are zero, and so these correspond to the offset and scale used to
position the raster - in combination with the number of rows and columns of data they
provide the spatial extent and the pixel size in each direction.
}
\examples{
f <- system.file("extdata", "sst.tif", package = "vapour")
raster_info(f)
}
