% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{raster_io}
\alias{raster_io}
\title{Raster IO (read)}
\usage{
raster_io(filename, window, band = 1L)
}
\arguments{
\item{filename}{data source}

\item{window}{src_offset, src_dim, out_dim}

\item{band}{index of which band to read}
}
\description{
Read a window of data from a GDAL raster source. The first argument is the source
name and the second is a 6-element \code{window} of offset, source dimension, and output dimension.
}
\details{
This is analgous to the \code{rgdal} function \code{readGDAL} with its arguments \code{offset},  \code{region.dim}
and \code{output.dim}.
}
\examples{
f <- system.file("extdata", "sst.tif", package = "vapour")
## a 5*5 window from a 10*10 region
raster_io(f, window = c(0, 0, 10, 10, 5, 5))
## find the information first
#ri <- raster_info(f)
#str(matrix(raster_io(f, c(0, 0, ri$dimXY, ri$dimXY)), ri$dimXY[1]))
## the method can be used to up-sample as well
#str(matrix(raster_io(f, window = c(0, 0, 10, 10, 15, 25)), 15))
## a future version will provide access to different methods
}
