% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gdal_attributes.R
\name{read_gdal_table}
\alias{read_gdal_table}
\title{Read attributes from source as a data frame}
\usage{
read_gdal_table(dsource, layer = 0, sql = NULL)
}
\arguments{
\item{dsource}{data source string}

\item{layer}{layer index (0)}

\item{sql}{optional sql query executed against the data source}
}
\description{
See here for SQL details http://www.gdal.org/ogr_sql.html
}
\examples{
f <- system.file("extdata", "point.shp", package = "vapour")
read_gdal_table(f, sql = "SELECT * FROM point WHERE FID > 4 AND FID < 8")
sfile <- system.file("shape/nc.shp", package="sf")
read_gdal_table(sfile, sql = "SELECT NAME, AREA, SID79 * 10 FROM nc WHERE PERIMETER < 1.2")
read_gdal_table(sfile)
}
