% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gdal_attributes.R
\name{read_gdal_table}
\alias{read_gdal_table}
\title{Read attributes from source as a data frame}
\usage{
read_gdal_table(dsource, layer = 0)
}
\arguments{
\item{dsource}{data source string}

\item{layer}{layer index (0)}
}
\description{
Read attributes from source as a data frame
}
\examples{
f <- system.file("extdata", "point.shp", package = "vapour")
read_gdal_table(f)
sfile <- system.file("shape/nc.shp", package="sf")
#vapour(sfile)
read_gdal_table(sfile)
}
