% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_bblob}
\alias{to_bblob}
\title{Read GDAL geometry bblob}
\usage{
to_bblob(dsource, layer = 0L)
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}
}
\description{
Read GDAL geometry bblob
}
\examples{
sfile <- system.file("shape/nc.shp", package="sf")
to_bblob(sfile)
}
