% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_binary}
\alias{to_binary}
\title{Read GDAL geometry as blob}
\format{indicate text output format, available are "json" (default), "gml", "kml", "wkt"}
\usage{
to_binary(dsource, layer = 0L)
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}
}
\description{
Simple read of geometry-only as WKB format.
}
\examples{
sfile <- system.file("shape/nc.shp", package="sf")
tib <- tibble::tibble(wkb = to_binary(sfile)) \%>\% bind_cols(read_gdal_table(sfile))
pfile <- "inst/extdata/point.shp"
to_binary(pfile)
}
