% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_binary}
\alias{to_binary}
\title{Read GDAL geometry as text}
\format{indicate text output format, available are "json" (default), "gml", "kml", "wkt"}
\usage{
to_binary(dsource, layer = 0L)
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}
}
\description{
Simple read of geometry-only as text format.
}
\details{
Microprocessors, databases, servers.
}
\examples{
sfile <- system.file("shape/nc.shp", package="sf")
vgeom(sfile)
pfile <- "inst/extdata/point.shp"
vgeom(pfile)
}
