% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy-deprecate.R
\name{vapour-deprecated}
\alias{vapour-deprecated}
\alias{vapour_read_geometry_what}
\alias{vapour_read_feature_what}
\alias{vapour_read_geometry_cpp}
\alias{sds_info}
\title{Functions deprecated from vapour}
\usage{
vapour_read_geometry_what(dsource, layer = 0L, sql = "",
  what = "geometry", textformat = "")

vapour_read_feature_what(dsource, layer = 0L, sql = "", what = "geometry",
  textformat = "json")

vapour_read_geometry_cpp(dsource, layer = 0L, sql = "", what = "geometry",
  textformat = "json")

sds_info(x)
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}

\item{sql}{if not empty this is executed against the data source (layer will be ignored)}

\item{textformat}{indicate text output format, available are "json" (default), "gml", "kml", "wkt"}

\item{x}{data source name}
}
\description{
These will be removed in a future release.
}
\seealso{
vapour_read_geometry_cpp

vapour_read_geometry_cpp

raster_sds_info
}
