% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_input.R
\name{vapour_layer_names}
\alias{vapour_layer_names}
\title{Layer names}
\usage{
vapour_layer_names(dsource, sql = "")
}
\value{
character vector of layer names
}
\description{
Obtain the names of available layers from a GDAL vector source.
}
\examples{
file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
vapour_layer_names(mvfile)
}
