% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vapour_read_attributes}
\alias{vapour_read_attributes}
\title{Vector attributes read}
\usage{
vapour_read_attributes(dsource, layer = 0L, sql = "")
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}

\item{sql}{if not empty this is executed against the data source (layer will be ignored)}
}
\description{
Read layer attributes, optionally after SQL select
}
\examples{
sfile <- system.file("shape/nc.shp", package="sf")
vapour_read_attributes(sfile)
sfile <- system.file("shape/nc.shp", package="sf")
vapour_read_attributes(sfile, "SELECT * FROM nc WHERE FID < 5")

pfile <- "inst/extdata/point.shp"
vapour(pfile)
}
