% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/vapour_input.R
\name{vapour_read_geometry}
\alias{vapour_read_geometry}
\alias{vapour_read_geometry_cpp}
\alias{vapour_read_geometry}
\alias{vapour_read_geometry_text}
\alias{vapour_read_geometry_cpp}
\alias{vapour_read_geometry}
\alias{vapour_read_geometry_text}
\title{Read GDAL feature geometry}
\usage{
vapour_read_geometry_cpp(dsource, layer = 0L, sql = "", what = "geometry",
  textformat = "json")

vapour_read_geometry(dsource, layer = 0L, sql = "")

vapour_read_geometry_text(dsource, layer = 0L, sql = "",
  textformat = "json")
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}

\item{sql}{if not empty this is executed against the data source (layer will be ignored)}

\item{what}{what to read, "geometry", "text", "extent"}

\item{textformat}{indicate text output format, available are "json" (default), "gml", "kml", "wkt"}
}
\description{
Read GDAL geometry as binary blob, text, or numeric extent.
}
\details{
\code{vapour_read_geometry} will read features as binary WKB, \code{vapour_read_geometry_text} as various text formats (geo-json, wkt, kml, gml),
\code{vapour_read_extent} a numeric extent which is the native bounding box, the four numbers (in this order) \code{xmin, xmax, ymin, ymax}.
For each function an optional SQL string will be evaluated against the data source before reading.

\code{vapour_read_geometry_cpp} will read a feature for each of the ways listed above and is used by those functions. It's recommended
to use the more specialist functions rather than this more general one.
}
\examples{
file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
#tib <- tibble::tibble(wkb = vapour_read_geometry(mvfile)) \%>\%
#  bind_cols(read_gdal_table(mvfile))
pfile <- system.file("extdata/point.shp", package = "vapour")
vapour_read_geometry(pfile)

file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
vapour_read_geometry_text(mvfile)
pfile <- system.file("extdata/point.shp", package = "vapour")
vapour_read_geometry_text(pfile)

file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
vapour_read_extent(mvfile)

file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
#tib <- tibble::tibble(wkb = vapour_read_geometry(mvfile)) \%>\%
#  bind_cols(read_gdal_table(mvfile))
pfile <- system.file("extdata/point.shp", package = "vapour")
vapour_read_geometry(pfile)

file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
tx_post_json <- vapour_read_geometry_text(mvfile)
pfile <- system.file("extdata/point.shp", package = "vapour")
tx_point_json <- vapour_read_geometry_text(pfile)
}
