% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vapour_read_geometry}
\alias{vapour_read_geometry}
\title{Read GDAL geometry as blob}
\format{indicate text output format, available are "json" (default), "gml", "kml", "wkt"}
\usage{
vapour_read_geometry(dsource, layer = 0L, sql = "")
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}

\item{sql}{if not empty this is executed against the data source (layer will be ignored)}
}
\description{
Simple read of geometry-only as WKB format.
}
\examples{
file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
tib <- tibble::tibble(wkb = vapour_read_geometry(mvfile)) \%>\%
  bind_cols(read_gdal_table(mvfile))
pfile <- "inst/extdata/point.shp"
vapour_read_geometry(pfile)
}
