% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vapour_read_geometry_text}
\alias{vapour_read_geometry_text}
\title{Read GDAL geometry as text}
\format{indicate text output format, available are "json" (default), "gml", "kml", "wkt"}
\usage{
vapour_read_geometry_text(dsource, layer = 0L, sql = "", format = "json")
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}
}
\description{
Simple read of geometry-only as text format.
}
\details{
Microprocessors, databases, servers.
}
\examples{
sfile <- system.file("shape/nc.shp", package="sf")
to_format(sfile)
pfile <- "inst/extdata/point.shp"
to_format(pfile)
}
