% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vapour_read_geometry_text}
\alias{vapour_read_geometry_text}
\title{Read GDAL geometry as text}
\usage{
vapour_read_geometry_text(dsource, layer = 0L, sql = "", format = "json")
}
\arguments{
\item{dsource}{data source name (path to file, connection string, URL)}

\item{layer}{integer of layer to work with, defaults to the first (0)}

\item{sql}{if not empty this is executed against the data source (layer will be ignored)}

\item{format}{indicate text output format, available are "json" (default), "gml", "kml", "wkt"}
}
\description{
Simple read of geometry-only as text format.
}
\examples{
file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
vapour_read_geometry_text(mvfile)
pfile <- system.file("extdata/point.shp", package = "vapour")
vapour_read_geometry_text(pfile)
}
