% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour-helpers.R
\name{vapour_srs_wkt}
\alias{vapour_srs_wkt}
\title{PROJ4 string to WKT}
\usage{
vapour_srs_wkt(crs)
}
\arguments{
\item{crs}{PROJ4 string}
}
\description{
Convert a PROJ4 string to Well Known Text.
}
\details{
The function is vectorized because why not, but probably only ever will be
used on single element vectors of character strings.
}
\examples{
vapour_srs_wkt("+proj=laea +datum=WGS84")
}
