% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_vrt.R
\name{vapour_vrt}
\alias{vapour_vrt}
\title{Virtual raster}
\usage{
vapour_vrt(dsn, extent = NULL, projection = NULL, bands = NULL, sds = 1L)
}
\arguments{
\item{dsn}{data source name, filepath, url, database connnection string, or VRT text}

\item{extent}{(optional) numeric extent, xmin,xmax,ymin,ymax}

\item{projection}{(optional) character string, projection string ("auth:code", proj4, or WKT, or anything understood by PROJ)}

\item{bands}{(optional) which band/s to include from the source}

\item{sds}{which subdataset to select from a source with more than one}
}
\value{

}
\description{
Simple VRT creation of a GDAL virtual raster.
}
\details{
Create a GDAL data source string (to be used like a filename) with various helpers.
VRT stands for 'ViRTual'.

An input string will be converted to a single subdataset, use 'sds' argument to select.

If 'extent', 'projection' is provided this is applied to override the source's extent and/or projection.
(These might be invalid, or missing, so we facilitate correcting this).

If 'bands' is provided this is used to select a set of bands (numbered from 1), which might be repeated, or
in any order.
}
\examples{
tif <- system.file("extdata", "sst.tif", package = "vapour")
vapour_vrt(tif)

vapour_vrt(tif, bands = c(1, 1))
}
